% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_data.R
\name{check_data}
\alias{check_data}
\title{Check overlap cohorts over the years
In the following matrix, 0 means the tables have no data in that year, and 1 means they have data in corresponding years.}
\usage{
check_data(config = "./phenotypes.json", db_file = "../nhanes.sqlite")
}
\arguments{
\item{config}{configuration file path}

\item{db_file}{data file path}
}
\value{
show the matrix
}
\description{
Check overlap cohorts over the years
In the following matrix, 0 means the tables have no data in that year, and 1 means they have data in corresponding years.
}
\examples{
check_data(config="./phenotypes.json",db_file="../nhanes.sqlite")
}
