% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_data.R
\name{checkDataConsistency}
\alias{checkDataConsistency}
\title{Check Variable Consistency}
\usage{
checkDataConsistency(table1, table2)
}
\arguments{
\item{table1}{NHANES table name 1}

\item{table2}{NHANES table name 2}
}
\value{
A matrix with columns named using the union of the column names for the two tables and three rows.
}
\description{
The function extracts both tables from the database and determines the union of their column names. A matrix with three rows and one column for each entry in the union is returned.  The first row indicates whether the variable was found in the first table and similarly the second row indicates whether the variable was found in the second table.  The third row of the matrix indicates whether or not the same encoding was used in both tables.  Users will typically want to use this function before merging tables.
}
\details{
Compare variables between two tables and report whether they are encoded the same.
}
\examples{
checkDataConsistency("DEMO_C","DEMO_D")
}
