% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-insert.R
\name{dbInsertNhanesTable}
\alias{dbInsertNhanesTable}
\title{Insert NHANES table into a database}
\usage{
dbInsertNhanesTable(
  con,
  x,
  data = nhanes(x, translated = FALSE),
  type = c("raw", "translated", "both"),
  codebook = NULL,
  make_primary_key = TRUE,
  check_integer = TRUE,
  cleanse_numeric = TRUE
)
}
\arguments{
\item{con}{An open database connection}

\item{x}{Character giving the name of an NHANES table}

\item{data}{Data frame. If unspecified, obtained from \code{x}
using \code{\link[pkg:nhanesA]{nhanes}}.}

\item{type}{Character, whether to insert raw data, translated, or
both.}

\item{codebook}{The variable codebooks for this table, as returned
by \code{\link[pkg:nhanesA]{nhanesCodebook}}}

\item{make_primary_key}{Logical flag, whether suitable variables
(usually \code{SEQN}) should be declared as primary keys.}

\item{check_integer}{Logical flag, whether numeric variables that
are all whole numbers should be stored as integer rather than
float / double.}

\item{cleanse_numeric}{Logical flag, indicating whether some some
special values for numeric variables should be converted to
missing values. TODO more details.}
}
\description{
Inserts an NHANES table into a database using the DBI
interface. Requires and active database connection, and assumes
that a table with this name does not already exist.
}
\author{
Deepayan Sarkar
}
