% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-insert.R
\name{addPrimaryKey}
\alias{addPrimaryKey}
\alias{insertTableDB}
\title{Declare one or more non-null columns to be primary key}
\usage{
addPrimaryKey(con, table, columns)

insertTableDB(
  con,
  data,
  table,
  check_integer = TRUE,
  non_null = NULL,
  make_primary_key = FALSE
)
}
\arguments{
\item{con}{An open database connection}

\item{table}{Character giving the name of a table.}

\item{columns}{Character vector, giving names of one or more columns in the table}

\item{data}{}

\item{check_integer}{}

\item{non_null}{}

\item{make_primary_key}{}
}
\description{
Declare one or more columns in an existing database table to be
primary key(s). Requires that the corresponding variables be
already defined as non-null.

General purpose DB table insertion
}
\details{
Try to insert an R data frame into an existing database with a
user-specified table name. Assumes that a table with this name
does not already exist.
}
\author{
Deepayan Sarkar
}
