% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_data.R
\name{jointQuery}
\alias{jointQuery}
\title{Joint Query}
\usage{
jointQuery(tables_n_cols, translated = TRUE)
}
\arguments{
\item{tables_n_cols}{a named list, each name corresponds to a Questionnaire and the value is a list of variable names.}

\item{translated}{whether the variables are translated}
}
\value{
This function returns a data frame containing the joined data from the specified tables and selected variables.
}
\description{
The jointQuery function is designed for merging and joining tables from different surveys over various years and returning the resultant data frame.
The primary objective of this function is to union given variables and tables from the same survey across different years, join different surveys, and return a unified data frame.
}
\examples{
df = jointQuery( list(BPQ_J=c("BPQ020", "BPQ050A"), DEMO_J=c("RIDAGEYR","RIAGENDR")))
cols = list(DEMO_I=c("RIDAGEYR","RIAGENDR","RIDRETH1","DMDEDUC2"),
                     DEMO_J=c("RIDAGEYR","RIAGENDR","RIDRETH1","DMDEDUC2"),
                     BPQ_I=c('BPQ050A','BPQ020'),BPQ_J=c('BPQ050A','BPQ020'),
                     HDL_I=c("LBDHDD"),HDL_J=c("LBDHDD"), TRIGLY_I=c("LBXTR","LBDLDL"),
                     TRIGLY_J=c("LBXTR","LBDLDL"))
ans = jointQuery(cols)
dim(ans)
}
