% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual.R
\name{make_bins}
\alias{make_bins}
\title{Make Bins for y values based on x}
\usage{
make_bins(x, y, nbin)
}
\arguments{
\item{x}{x values}

\item{y}{y values}

\item{nbin}{expected number of data point in each bin}
}
\value{
a data frame with binned data frame
}
\description{
Make Bins for y values based on x
}
\examples{
make_bins(x,y,200)
}
