% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metaData}
\alias{metaData}
\title{Show meta data information}
\usage{
metaData(meta)
}
\arguments{
\item{meta}{meta values. If meta set as \code{T} or \code{Tables}, it shows the information of NHANES tables.
If meta set as \code{V} or \code{Variables}, it shows the information of NHANES Variables.
If meta set as \code{C} or \code{Codebook}, it shows the Codebook information.}
}
\value{
dataframe contains the meta data
}
\description{
It shows NAHANES medata such as, tables, variables and codebooks.
}
\examples{
metaData("T")
metaData("Codebook")
}
