% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-codebook.R
\name{metadata_cb}
\alias{metadata_cb}
\alias{metadata_tab}
\alias{metadata_var}
\title{Metadata Tables : Access Postgres DB}
\usage{
metadata_cb(variable = NULL, table = NULL)

metadata_var(variable = NULL, table = NULL)

metadata_tab(table = NULL, UseConstraints = FALSE)
}
\arguments{
\item{variable}{Character vector naming one or more variable}

\item{table}{Character vector naming one or more NHANES table}

\item{UseConstraints}{If FALSE then only tables with no use
constraints are returned.}
}
\value{
A dataframe or tibble with the appropriate subset of the
metadata table.
}
\description{
Access Metadata Tables in NHANES Postgres DB
}
\details{
\code{metadata_cb} retrieves data from the VariableCodebook table,
\code{metadata_var} retrieves data from the QuestionnaireVariables table
and \code{metadata_tab} retrieves data from the QuestionnaireDescriptions table.
Where appropriate the returned value can be a subset for a single variable, or
NHANES table. The contents of these tables is described in the MetaData vignette for the
package.

metadata_var accesses the QuestionnaireVariables metadata
table.  metadata_cb accesses the VariableCodebook metadata
table.  metadata_tab accesses the QuestionnaireDescriptions
table.  The returned object has entries for each
variable/table combination where there is a match, in the
sense that the table has a variable with the supplied name.
}
\examples{
ex1 = metadata_cb(variable = "LBDLDL")
ex2 = metadata_var(table = "DEMO_D")
ex3 = metadata_tab(table = "ACQ_J")
ex4 = metadata_var(variable=c("RIDAGEYR",  "RIAGENDR", "BPXSY1"), table=c("DEMO", "DEMO_B", "DEMO_C", "BPX_C"))
}
\author{
Deepayan Sarkar
}
