% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanesA.R
\name{nhanes}
\alias{nhanes}
\title{Download/Load an NHANES table and return as a data frame.}
\usage{
nhanes(nh_table)
}
\arguments{
\item{nh_table}{The name of the specific table to retrieve.}
}
\value{
data frame
}
\description{
Use to download/load NHANES data tables that are in SAS format.
}
\examples{
nhanes('BPX_E')
}
