% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanesA.R
\name{nhanesCDCDataDoc}
\alias{nhanesCDCDataDoc}
\title{Show the doc page}
\usage{
nhanesCDCDataDoc(nh_table)
}
\arguments{
\item{nh_table}{a NHANES table name}
}
\value{
open the CDC doc page
}
\description{
Show the doc page
}
\examples{
nhanesCDCDataDoc("DEMO_C")
}
