% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanesA.R
\name{nhanesCodebook}
\alias{nhanesCodebook}
\title{Display codebook for selected variable.}
\usage{
nhanesCodebook(nh_table, colname)
}
\arguments{
\item{nh_table}{The name of the NHANES table that contains the desired variable.}

\item{colname}{the name of the table column (variable).}
}
\value{
The codebook is returned as a list object. Returns NULL upon error.
}
\description{
Display codebook for selected variable.
}
\examples{
t1 = nhanesCodebook('AUX_D', 'AUQ020D')
t2 = nhanesCodebook('DEMO_D', 'RIAGENDR')
}
