% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_data.R
\name{nhanesColnames}
\alias{nhanesColnames}
\title{Column Names for NHANES tables}
\usage{
nhanesColnames(tb_name)
}
\arguments{
\item{tb_name}{NHANES table name}
}
\value{
a character vector of non-zero length equal to the appropriate dimension.
}
\description{
Retrieve column names of an NHANES table.
}
\examples{
nhanesColnames("BMX_I")
}
