% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_data.R
\name{nhanesDim}
\alias{nhanesDim}
\title{Dimensions of NHANES table}
\usage{
nhanesDim(tb_name)
}
\arguments{
\item{tb_name}{NHANES table name}
}
\value{
retrieves the dim attribute of NHANES table.
}
\description{
Dimensions of NHANES table
}
\examples{
nhanesDim("BMX_I")
}
