% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_data.R
\name{nhanesHead}
\alias{nhanesHead}
\title{Return the First of an NHANES table}
\usage{
nhanesHead(nh_table, n = 5)
}
\arguments{
\item{nh_table}{NHANES table name}

\item{n}{number of rows of NHANES table}
}
\value{
retrieves the First of an NHANES table
}
\description{
Return the First of an NHANES table
}
\examples{
nhanesHead("BMX_I")
nhanesHead("BMX_I",10)
}
