% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_data.R
\name{nhanesNcol}
\alias{nhanesNcol}
\title{The Number of Columns of an NHANES table}
\usage{
nhanesNcol(tb_name, translated = TRUE)
}
\arguments{
\item{tb_name}{NHANES table name}

\item{translated}{whether the table name is translated}
}
\value{
an integer of length 1
}
\description{
The Number of Columns of an NHANES table
}
\examples{
nhanesNcol("BMX_I")
}
