% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_data.R
\name{nhanesNrow}
\alias{nhanesNrow}
\title{The Number of Rows of an NHANES table}
\usage{
nhanesNrow(tb_name)
}
\arguments{
\item{tb_name}{NHANES table name}
}
\value{
an integer of length 1
}
\description{
The Number of Rows of an NHANES table
}
\examples{
nhanesNrow("BMX_I")
}
