% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanesA.R
\name{nhanesSearchVarName}
\alias{nhanesSearchVarName}
\title{Search for tables that contain a specified variable,implementation of nhanesA::nhanesSearchVarName()}
\usage{
nhanesSearchVarName(
  varnames = NULL,
  ystart = NULL,
  ystop = NULL,
  includerdc = FALSE,
  nchar = 128,
  namesonly = TRUE
)
}
\arguments{
\item{varnames}{Names of variable to match.}

\item{ystart}{Four digit year of first survey included in search, where ystart >= 1999.}

\item{ystop}{Four digit year of final survey included in search, where ystop >= ystart.}

\item{includerdc}{If TRUE then RDC only tables are included in list (default=FALSE).}

\item{nchar}{Truncates the variable description to a max length of nchar.}

\item{namesonly}{If TRUE then only the table names are returned (default=TRUE).}
}
\value{
By default, a character vector of table names that include the specified variable is returned. If namesonly=FALSE, then a data frame of table attributes is returned.
}
\description{
Search for tables that contain a specified variable,implementation of nhanesA::nhanesSearchVarName()
}
\details{
The NHANES Comprehensive Variable List is scanned to find all data tables that contain the given variable name. Only a single variable name may be entered, and only exact matches will be found.
}
\examples{
nhanesSearchVarName(c('BPXPULS','BMXBMI'))
nhanesSearchVarName(c('BPXPULS','BMXBMI'),ystop=2004)
}
