% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanesA.R
\name{nhanesTableVars}
\alias{nhanesTableVars}
\title{Displays a list of variables in the specified NHANES table.}
\usage{
nhanesTableVars(
  data_group,
  nh_table,
  details = FALSE,
  nchar = 128,
  namesonly = FALSE
)
}
\arguments{
\item{data_group}{data_group: The type of survey (DEMOGRAPHICS, DIETARY, EXAMINATION,LABORATORY, QUESTIONNAIRE). Abbreviated terms may also beused: (DEMO, DIET, EXAM, LAB, Q).}

\item{details}{If TRUE then a more detailed description of the tables is returned (default=FALSE).}

\item{nchar}{Truncates the table description to a max length of nchar.}

\item{namesonly}{If TRUE then only the table names are returned (default=FALSE).}

\item{nh_tableThe}{name of the specific table to retrieve.}
}
\value{
The number of characters in the Variable Description to print. Default length is 128, which is set to enhance readability cause variable descriptions can be very long.
}
\description{
Displays a list of variables in the specified NHANES table.
}
\examples{
nhanesTableVars('LAB', 'CBC_E')
}
