% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_data.R
\name{nhanesTail}
\alias{nhanesTail}
\title{Return the Last of an NHANES table}
\usage{
nhanesTail(nh_table, n = 5, translated = TRUE)
}
\arguments{
\item{nh_table}{NHANES table name}

\item{n}{number of rows of NHANES table}

\item{translated}{whether the table name is translated}
}
\value{
retrieves the Last of an NHANES table
}
\description{
Return the Last of an NHANES table
}
\examples{
nhanesTail("BMX_I")
nhanesTail("BMX_I",10)
}
