% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanesA.R
\name{nhanesTranslate}
\alias{nhanesTranslate}
\title{Display code translation information.}
\usage{
nhanesTranslate(
  nh_table,
  colnames = NULL,
  data = FALSE,
  nchar = 32,
  mincategories = 2,
  details = FALSE,
  dxa = FALSE
)
}
\arguments{
\item{nh_table}{The name of the NHANES table to retrieve.}

\item{colnames}{The names of the columns to translate.}

\item{data}{If TRUE then the table will be translated and a a dataframe returned.}

\item{nchar}{Applies only when data is defined. Code translations can be very long. Truncate the length by setting nchar (default = 32).}

\item{mincategories}{The minimum number of categories needed for code translations to be applied to the data (default=2).}

\item{details}{If TRUE then all available table translation information is displayed (default=FALSE).}

\item{dxa}{If TRUE then the 2005-2006 DXA translation table will be used (default=FALSE).}
}
\value{
The code translation table (or translated data frame when data is defined). Returns NULL upon error.
}
\description{
Display code translation information.
}
\details{
Most NHANES data tables have encoded values. E.g. 1 = 'Male', 2 = 'Female'. Thus it is often helpful to view the code translations and perhaps insert the translated values in a data frame. Only a single table may be specified, but multiple variables within that table can be selected. Code translations are retrieved for each variable.
}
\examples{
nhanesTranslate("DEMO_C",c("RIAGENDR","RIDRETH1"))
nhanesTranslate("DEMO_C",c("RIAGENDR","RIDRETH1"),data=TRUE)
data[1:10,]
}
