% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nonPhenotypes}
\alias{nonPhenotypes}
\title{non phenotypes data}
\format{
\subsection{\code{nonPhenotypes}}{

A data frame with 515 rows and 3 columns:
\describe{
\item{names}{Country name}
\item{types}{type of the  data}
\item{tables}{nhanes tabless}
}
}
}
\source{
{created manully}
}
\usage{
nonPhenotypes
}
\description{
non phenotypes data
}
\examples{
 data(nonPhenotypes)
}
\keyword{datasets}
