% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phesant.R
\name{phesant}
\alias{phesant}
\title{simple version of PHEnome Scan ANalysis Tool (PHESANT)}
\usage{
phesant(df)
}
\arguments{
\item{df}{data frame to be processed}
}
\value{
suggested data types
}
\description{
simple version of PHEnome Scan ANalysis Tool (PHESANT)
}
\examples{
phesant(df=nhanes('DEMO_C'))
}
