% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_data.R
\name{phseant_table}
\alias{phseant_table}
\title{Create data frame shows the variable}
\usage{
phseant_table(phs_types, dsc_config)
}
\arguments{
\item{phs_types}{the data types assigned to phenotype by the phseant function}

\item{dsc_config}{search tables}
}
\value{
results data frame
}
\description{
Create data frame shows the variable
}
\examples{
function(phs_types,dsc_config)
}
