% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_keys.R
\name{primary_keys}
\alias{primary_keys}
\title{primary_keys: Primary keys for NHANES table}
\usage{
primary_keys(x, require_unique = FALSE)
}
\arguments{
\item{x}{Character string giving name of an NHANES table}

\item{require_unique}{Logical; whether the likely intended primary
key variables should be returned even if they do not uniquely
identify rows.}
}
\value{
Character vector giving variables that should serve as
primary keys for the table. May be \code{NULL} if
\code{require_unique = TRUE}.
}
\description{
Look up primary keys for a given NHANES table
}
\details{
Most NHANES tables contain a variables called SEQN that represents
a participant ID, and hence can be used as a primary key for
join operations. Some tables are in the long format, potentially
containing multiple observations per participant, where other
variables along with SEQN serve as a composite primary key. Yet
other tables have no participant ID, where the primary key
variable depends on the specific context. For some tables
(mostly those containing dietary components codes and drug /
supplement codes) the intended primary key columns are not
actually unique.
}
\author{
Deepayan Sarkar
}
