% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-codebook.R
\name{qc_var}
\alias{qc_var}
\title{QC report for a variable in NHANES}
\usage{
qc_var(x, var = metadata_var(x), cb = metadata_cb(x), tab = metadata_tab())
}
\arguments{
\item{x}{}

\item{var}{Optional data frame containing variable metadata}

\item{cb}{Optional data frame containing codebook metadata}

\item{tab}{Optional data frame containing table metadata}
}
\description{
.. content for \details{} ..
}
\author{
Deepayan Sarkar
}
