% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_util.R
\name{query}
\alias{query}
\title{Query data from the Docker database}
\usage{
query(sql)
}
\arguments{
\item{sql}{string of sql}
}
\value{
a data frame of the results
}
\description{
Query data from the Docker database
}
\examples{
query("SELECT TOP(50) * FROM QuestionnaireVariables;")
}
