% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_data.R
\name{query_data}
\alias{query_data}
\title{Query data with tables and columns provided by the configuration file}
\usage{
query_data(
  config = "./phenotypes.json",
  db_file = "./nhanes.sqlite",
  na.keep = TRUE
)
}
\arguments{
\item{config}{configuration file path}

\item{db_file}{data file path}

\item{na.keep}{all rows with NA values are remove when set as FALSE,}
}
\value{
data
}
\description{
Query data with tables and columns provided by the configuration file
}
\examples{
df <- query_joint_data("./phenotypes.json","../nhanes.sqlite")
}
