% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_config.R
\name{read_conf}
\alias{read_conf}
\title{Read configuration information for the table}
\usage{
read_conf(table_name = "DEMO_I")
}
\arguments{
\item{table_name}{table name}
}
\value{
parsed json file contented or NULL
}
\description{
Read configuration information for the table
}
\examples{
read_conf("DEMO_I")
read_conf(table_name="DEMO_I")
}
