% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_data.R
\name{unionQuery}
\alias{unionQuery}
\title{Query Data with names of the tables}
\usage{
unionQuery(tb_names, cols = NULL)
}
\arguments{
\item{tb_names}{names of the tables}

\item{cols}{columns}
}
\value{
queried data frane
}
\description{
When multiple tables have been assigned, it will query all the tables, combine them, and only keeps the columns shared by all the tables.
}
\examples{
unionQuery(tb_names=c("BPX_D","BPX_E"),cols=c("BPXDI1","BPXDI2"))
unionQuery(tb_names=c("PhthalatesPhytoestrogensAndPAHsUrine","PhytoestrogensUrine"),cols=c("URXDAZ","URXDMA"))

}
