% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{variableMetaData}
\alias{variableMetaData}
\title{Retrieve variable meta data information from Metadata.QuestionnaireVariables}
\usage{
variableMetaData(varList)
}
\arguments{
\item{varList:}{This is a named list, the names are valid NHANES table/questionnaire names
and the elements are the names of variables.}
}
\value{
A named list of dataframes, the names correspond to the names from the input list
and the values are dataframes that contain the information in the corresponding rows
of the Metadata.QuestionnaireVariables table
}
\description{
Retrieve variable meta data information from Metadata.QuestionnaireVariables
}
\examples{
input = list(AA_H = c("SEQN", "WTSA2YR", "URX2NP"),
    DEMO=c("SEQN", "RIDAGEYR","RIAGENDR","RIDRETH1","DMDEDUC2", "INDFMPIR"))
ans = variableMetaData(input)
ans[[1]][1,]
}
