% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_data.R
\name{checkDataConsistency}
\alias{checkDataConsistency}
\title{Check Variable Consistency}
\usage{
checkDataConsistency(table1, table2)
}
\arguments{
\item{table1}{NHANES table name 1}

\item{table2}{NHANES table name 2}
}
\value{
it returns a matrix, and the first and second row show whether the variable existing in table 1 or table 2, TRUE values means existing, FALSE means not existing. The third whether the variables are encoded in the same values if they shown in both tables, TRUE means encoded as the same, FALSE means they are encoded as the different value.
}
\description{
Check if the variables across over two tables and encoded as the same values.
}
\examples{
checkDataConsistency("DEMO_C","DEMO_D")
}
