% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_data.R
\name{dataDescription}
\alias{dataDescription}
\title{Get the metadata associated with a set of tables and variables}
\usage{
dataDescription(tables_n_cols)
}
\arguments{
\item{tables_n_cols}{a named list, each name corresponds to a Questionnaire and the value is a list of variable names.}
}
\value{
A dataframe where the columns are various metadata values and the rows correspond to the input variable names.
}
\description{
Get the metadata associated with a set of tables and variables
}
