% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-skip.R
\name{get_skip_info}
\alias{get_skip_info}
\title{Find variables that may have been skipped based on response to a
previous question.}
\usage{
get_skip_info(table)
}
\arguments{
\item{table}{Name of a NHANES table}
}
\value{
data frame
}
\description{
Find variables that may have been skipped based on response to a
previous question.
}
\examples{
sk1 = get_skip_info("DEMO_J")
table(sk1$MaybeSkipped)
sk1[sk1$MaybeSkipped,]
}
\author{
Deepayan Sarkar
}
