% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-skip.R
\name{get_skip_info}
\alias{get_skip_info}
\title{Identify questions that may have been skipped based on response to a
previous question.}
\usage{
get_skip_info(table)
}
\arguments{
\item{table}{Character string giving the name of an NHANES table}
}
\value{
A data frame with columns:
\describe{
\item{Table}{The name of the input table.}
\item{Variable}{One row for each question in the documentation file.}
\item{MaybeSkipped}{A logical value indicating whether that question might have been skipped.}
.  \item{SkippedDueTo}{A vector of the questions, delivered earlier, that might have caused this question to be skipped.}
}
}
\description{
Identify questions that may have been skipped based on response to a
previous question.
}
\examples{
sk1 = get_skip_info("DEMO_J")
table(sk1$MaybeSkipped)
sk1[sk1$MaybeSkipped,]
}
\author{
Deepayan Sarkar
}
