% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_data.R
\name{jointQuery}
\alias{jointQuery}
\title{Joint Query}
\usage{
jointQuery(patterns, cols = NULL)
}
\arguments{
\item{cols}{}
}
\description{
Joint Query
}
\examples{
jointQuery(c("DEMO","BMX"))
patterns = c("DEMO","BMX")
cols = c("RIDAGEYR","RIAGENDR","BMXBMI","DMDEDUC2")
jointQuery(patterns,cols)
}
