% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_data.R
\name{jointQuery}
\alias{jointQuery}
\title{Joint Query}
\usage{
jointQuery(table_names, cols = NULL)
}
\arguments{
\item{table_names}{list of the table names want to joint and query}

\item{cols}{columns}
}
\value{
it merges the researched results and returns the results as a data frame.
}
\description{
Joint Query
}
\examples{
jointQuery(c("DEMO","BMX"))
jointQuery(c("DEMO","BodyMeasures"))
}
