% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metaData}
\alias{metaData}
\title{Retrieve meta data information from the SQL database}
\usage{
metaData(meta)
}
\arguments{
\item{meta}{meta values. If meta set as \code{T} or \code{Tables}, it shows the information on NHANES tables.
If meta set as \code{V} or \code{Variables}, it returns the information of NHANES Variables.
If meta set as \code{C} or \code{Codebook}, it returns the Codebook information.}
}
\value{
dataframe containing the meta data
}
\description{
Retrieve NHANES medata tables, variables and codebooks.
}
\examples{
metaData("T")
metaData("Codebook")
}
