% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanesA.R
\name{nhanesCodebook}
\alias{nhanesCodebook}
\title{Display the NHANES codebook for selected variable.}
\usage{
nhanesCodebook(nh_table, colname)
}
\arguments{
\item{nh_table}{The name of the NHANES Questionaire that contains the desired variable.}

\item{colname}{The NHANES variable name.}
}
\value{
The codebook is returned as a list with components
Variable, Description, Target, SasLabel and Codebook
}
\description{
Display the NHANES codebook for selected variable.
}
\examples{
t1 = nhanesCodebook('AUX_D', 'AUQ020D')
t2 = nhanesCodebook('DEMO_D', 'RIAGENDR')
}
