% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_data.R
\name{nhanesDim}
\alias{nhanesDim}
\title{Dimensions of NHANES table}
\usage{
nhanesDim(nh_table, translated = TRUE)
}
\arguments{
\item{nh_table}{NHANES table name}

\item{translated}{whether the table name is translated}
}
\value{
retrieves the dim attribute of NHANES table.
}
\description{
Dimensions of NHANES table
}
\examples{
nhanesDim("BMX_I")
}
