% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_data.R
\name{nhanesQuery}
\alias{nhanesQuery}
\title{Query NHANES data from Database}
\usage{
nhanesQuery(sql)
}
\arguments{
\item{sql}{query string for Microsoft SQL Server database.}
}
\value{
data frame
}
\description{
Query NHANES data from Database
}
\examples{
 demo = nhanesQuery("select * from Translated.DEMO_C")
 demo = nhanesQuery("select * from Raw.DEMO_C")
}
