% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanesA.R
\name{nhanesQuery}
\alias{nhanesQuery}
\title{Query data from the Docker database}
\usage{
nhanesQuery(sql)
}
\arguments{
\item{sql}{string of sql}
}
\value{
a data frame of the results
}
\description{
It is an internal function.
}
\examples{
query("SELECT TOP(50) * FROM QuestionnaireVariables;")
}
