% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanesA.R
\name{nhanesTables}
\alias{nhanesTables}
\title{Title}
\usage{
nhanesTables(
  data_group,
  year,
  nchar = 128,
  details = FALSE,
  namesonly = FALSE,
  includerdc = FALSE
)
}
\arguments{
\item{includerdc}{}
}
\description{
Title
}
