% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual.R
\name{plot_bins2}
\alias{plot_bins2}
\title{Plot multiple Bins}
\usage{
plot_bins2(df_list, xlab = "x", ylab = "Binned y", is_facet = TRUE)
}
\arguments{
\item{df_list}{a list contains data frames}

\item{xlab}{x-axis label}

\item{ylab}{x-axis label}

\item{is_facet}{Boolean flag to decide create facets}
}
\value{
plots binned data
}
\description{
Plot multiple Bins
}
\examples{
dflist = list( "norm"=make_bins(x=runif(600, 1, 60) ,y=rnorm(600),nbin=60),
                         "romm"=make_bins(x=runif(600, 1, 60) ,y=runif(600, -1, 1),nbin=60))
plot_bins2(dflist,is_facet = FALSE)
}
