% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-codebook.R
\name{qc_var}
\alias{qc_var}
\alias{summary.qc_var}
\alias{print.qc_var}
\title{qc_var: QC on NHANES variable}
\usage{
qc_var(
  x,
  tables = tables,
  var = metadata_var(x),
  cb = metadata_cb(x),
  tab = metadata_tab(),
  UseConstraints = FALSE
)

\method{summary}{qc_var}(object, ...)

\method{print}{qc_var}(x, ..., show = c("multiple", "description", "saslabel", "target"))
}
\arguments{
\item{x}{An object of class \code{"qv_var"}}

\item{var}{Optional data frame containing variable metadata}

\item{cb}{Optional data frame containing codebook metadata}

\item{tab}{Optional data frame containing table metadata}

\item{UseConstraints}{If FALSE, the default, then only publicly available tables are searched.}

\item{object}{An object of class \code{"qv_var"}}

\item{...}{Additional arguments, ignored}

\item{show}{A vector giving the types of identified anomalies to
print. Currently four types of anomalies are identified:
presence of the variable in multiple tables in the same cycle;
mismatches in the Description field, mismatches in SAS Label,
and mismatch in the Target specification. Some of these
mimatches are benign, so it is useful to suppress them when
printing.}
}
\value{
An object of S3 class \code{"qc_var"} with suitable print and summary methods.
}
\description{
QC report for a single variable in NHANES
}
\details{
The arguments var, cb and tab, will be default use the corresponding metadata tables (Variables, CodeBook and Questionnaires).
}
\examples{
t1 = qc_var("DMDEDUC3")
## restrict the tables
t2 = qc_var("DMDEDUC3", tables=c("DEMO_B", "DEMO_J"))
print(t2, show = "saslabel")
qc_var("LBCBHC")
qc_var("LBXHCT")
}
\author{
Deepayan Sarkar
}
