% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-codebook.R
\name{qc_var}
\alias{qc_var}
\alias{summary.qc_var}
\alias{print.qc_var}
\title{qc_var: QC on NHANES variable}
\usage{
qc_var(
  x,
  tables = tables,
  var = metadata_var(x),
  cb = metadata_cb(x),
  tab = metadata_tab(),
  UseConstraints = FALSE
)

\method{summary}{qc_var}(
  object,
  ...,
  show = c("multiple", "description", "saslabel", "target", "levels")
)

\method{print}{qc_var}(
  x,
  ...,
  show = c("multiple", "description", "saslabel", "target", "levels")
)
}
\arguments{
\item{x}{An object of class \code{"qc_var"}}

\item{tables}{Optional character vector of table names. If
specified, the search for the variable \code{x} is restricted
to these tables.}

\item{var}{Optional data frame containing variable metadata.}

\item{cb}{Optional data frame containing codebook metadata.}

\item{tab}{Optional data frame containing table metadata.}

\item{UseConstraints}{If FALSE, the default, then only publicly
available tables are searched.}

\item{object}{An object of class \code{"qc_var"}}

\item{...}{Additional arguments, ignored}

\item{show}{A vector giving the types of detected anomalies to
display. Currently five types of anomalies are identified:
presence of the variable in multiple tables in the same cycle;
mismatches in the Description field, mismatches in SAS Label,
mismatches in the Target specification, and inconsistencies in
levels. Some of these mimatches are benign, so it is useful to
suppress them when displaying the results.}
}
\value{
An object of S3 class \code{"qc_var"} with suitable print
and summary methods.
}
\description{
QC report for a single variable in NHANES
}
\details{
The arguments var, cb and tab, will be default use the
corresponding metadata tables (Variables, CodeBook and
Questionnaires).
}
\examples{
qc_var("LBCBHC")
qc_var("LBXHCT")
qc_var("RIDAGEYR") |> summary()
qc_var("DMDEDUC3") |> summary()
t2 = qc_var("DMDEDUC3", tables=c("DEMO_B", "DEMO_J"))
print(t2, show = "saslabel")
}
\author{
Deepayan Sarkar
}
