% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_data.R
\name{queryByVars}
\alias{queryByVars}
\title{Query data by variable}
\usage{
queryByVars(vars = NULL, ystart = NULL, ystop = NULL)
}
\arguments{
\item{vars}{variables or phenotypes want to search}

\item{ystart}{Four digit year of first survey included in search, where ystart >= 1999.}

\item{ystop}{Four digit year of final survey included in search, where ystop >= ystart.}
}
\value{
union data frame
}
\description{
It search the tables contain the given variables and query data union
}
\examples{
df = queryByVars(c("URXDAZ","URXDMA"))
}
