% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_data.R
\name{queryData}
\alias{queryData}
\title{Query Data with names of the tables}
\usage{
queryData(tb_names, cols = NULL)
}
\arguments{
\item{tb_names}{names of the tables}

\item{cols}{columns}
}
\value{
queried data
}
\description{
Query Data with names of the tables
}
