% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_data.R
\name{unionQuery}
\alias{unionQuery}
\title{Union Query}
\usage{
unionQuery(tables_n_cols, translated = TRUE)
}
\arguments{
\item{tables_n_cols}{a named list, each name corresponds to a Questionnaire and the value is a list of variable names.}

\item{translated}{A boolean parameter, default is TRUE. This indicates whether the variables are translated or not.}
}
\value{
data frame
}
\description{
The unionQuery function is used for merging or unifying tables from the same survey over different years and returning the resultant data frame.
The main goal of this function is to union given variables and tables from the same survey across different years.
}
\examples{
df = unionQuery( list(DEMO_I=c("RIDAGEYR","RIAGENDR"), DEMO_J=c("RIDAGEYR","RIAGENDR")))
}
