% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_data.R
\name{unionQuery}
\alias{unionQuery}
\title{Union Query}
\usage{
unionQuery(table_names, cols = NULL)
}
\arguments{
\item{table_names}{nhanes table names}

\item{cols}{columns}
}
\value{
data frame
}
\description{
Union Query
}
\examples{
df = unionQuery(c("DEMO_B","DEMO_D"),c("RIDAGEYR","RIAGENDR"))
}
